src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-ckeditor==6.5.1
django-colorfield==0.9.0
django-crispy-forms==2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-nested-routers==0.93.4
frozenlist==1.3.3
geoip2==4.7.0
idna==3.4
maxminddb==2.4.0
multidict==6.0.4
# numpy==1.24.3
# opencv-python==4.7.0.72
Pillow==9.5.0
psycopg2-binary==2.9.6
python-barcode==0.14.0
python-decouple==3.8
pytz==2023.3
pyzbar==0.1.9
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
